+++ src/protocols/mod.rs
@@ -1,12 +1,12 @@
-mod multicast;
 mod tcp;
-mod unicast;
+mod udp;
 
-pub use multicast::{MulticastReceiver, MulticastSender};
+use std::net::SocketAddr;
+
 pub use tcp::{TcpReceiver, TcpSender};
-pub use unicast::{UnicastReceiver, UnicastSender};
+pub use udp::{UdpReceiver, UdpSender};
 
-use anyhow::Result;
+use thiserror::Error;
 
 /// Buffer size for receive operations
 ///
@@ -14,9 +14,19 @@ use anyhow::Result;
 const RECV_BUF: usize = 65536;
 
 pub trait Sender {
-    fn send(&mut self, data: &[u8]) -> Result<()>;
+    fn send(&mut self, data: &[u8]) -> Result<(), ProtoError>;
 }
 
 pub trait Receiver {
-    fn recv<'a>(&mut self) -> Result<&[u8]>;
+    fn recv<'a>(&mut self) -> Result<&[u8], ProtoError>;
+}
+
+#[derive(Error, Debug)]
+pub enum ProtoError {
+    #[error("Client {0} disconnected")]
+    Disconnected(SocketAddr),
+    #[error("Client {0} is not connected: datagram ignored")]
+    ConflictingClient(SocketAddr),
+    #[error("IO Error: {0}")]
+    IOErr(#[from] std::io::Error),
 }