+++ src/protocols/tcp.rs
@@ -1,16 +1,14 @@
-use anyhow::{Error, Result};
-
 use std::io::{BufRead, BufReader, BufWriter, Write};
 use std::net::{Ipv4Addr, TcpListener, TcpStream};
 
-use super::{Receiver, Sender, RECV_BUF};
+use super::{ProtoError, Receiver, Sender, RECV_BUF};
 
 pub struct TcpSender {
     socket: BufWriter<TcpStream>,
 }
 
 impl TcpSender {
-    pub fn new(peer: Ipv4Addr, port: u16, _bind: Ipv4Addr) -> Result<Self> {
+    pub fn new(peer: Ipv4Addr, port: u16, _bind: Ipv4Addr) -> anyhow::Result<Self> {
         let socket = TcpStream::connect((peer, port))?;
         println!("Connected to server {peer}:{port}");
         Ok(Self {
@@ -20,13 +18,19 @@ impl TcpSender {
 }
 
 impl Sender for TcpSender {
-    fn send(&mut self, data: &[u8]) -> Result<()> {
+    fn send(&mut self, data: &[u8]) -> Result<(), ProtoError> {
         if let Err(e) = self.socket.write_all(data) {
-            println!("Disconnected from server. Reason: {e}");
-            Err(Error::from(e))
+            eprintln!("Disconnected from server. Reason: {e}");
+            Err(ProtoError::Disconnected(self.socket.get_ref().peer_addr()?))
         } else {
             Ok(())
         }
+    }
+}
+
+impl Drop for TcpSender {
+    fn drop(&mut self) {
+        let _x = self.socket.write_all(&[0]);
     }
 }
 
@@ -37,7 +41,7 @@ pub struct TcpReceiver {
 }
 
 impl Receiver for TcpReceiver {
-    fn recv<'a>(&mut self) -> Result<&[u8]> {
+    fn recv<'a>(&mut self) -> Result<&[u8], ProtoError> {
         let mut conn = if let Some(conn) = self.connection.take() {
             conn
         } else {
@@ -47,12 +51,13 @@ impl Receiver for TcpReceiver {
         };
         self.buf.clear();
         let res = conn.read_until(0, &mut self.buf);
-        self.connection = match &res {
+        match &res {
             Err(e) => {
                 println!("client disconnected: {}", e);
-                None
+                self.connection = None;
+                return Err(ProtoError::Disconnected(conn.get_ref().peer_addr()?));
             }
-            Ok(_) => Some(conn),
+            Ok(_) => self.connection = Some(conn),
         };
         res?;
         return Ok(&self.buf[..]);
@@ -60,7 +65,7 @@ impl Receiver for TcpReceiver {
 }
 
 impl TcpReceiver {
-    pub fn new(bind: Ipv4Addr, port: u16) -> Result<Self> {
+    pub fn new(bind: Ipv4Addr, port: u16) -> anyhow::Result<Self> {
         let listener = TcpListener::bind((bind, port))?;
 
         Ok(Self {