+++ src/speed_controllers/unlimited.rs
@@ -17,8 +17,6 @@ impl Limiter for UnLimiter {
     }
 }
 
-
-
 #[cfg(test)]
 mod tests {
     use super::*;